+++ graphics/gui_elements/gui_entities.cpp
@@ -39,6 +39,7 @@ void Gui::entities(Scene &scene, Config &c) {
         
         ImGui::Checkbox("Display collision boxes", &c.render_collision_boxes);
         ImGui::Checkbox("Use shaders", &c.use_shaders);
+        ImGui::SliderInt("Palette", &c.palette_index, 0, c.num_palettes - 1);
     }
     
     ImGui::End();