+++ graphics/renderer.h
@@ -14,7 +14,6 @@
 #include "ecs.h"
 #include "config.h"
 
-
 #define RENDER_SCALE 4
 
 namespace Verse
@@ -31,8 +30,10 @@ namespace Verse
         void render(Scene &scene, Config &c);
         void display(Config &c);
     
-        void presentBackBuffer(SDL_Renderer *renderer, SDL_Window* win, SDL_Texture* backBuffer, ui32 pid);
-        static ui32 programId;
+        void presentWithShaders(Config &c);
+        static ui8 programId;
+        static SDL_Texture *palette_tex;
+        static unsigned int textures[2];
     
         void calculateRefreshRate();
         static int refreshRate = 60;